Update data
This is to include a computation reported in the paper which was not
available in the Containerfile: we also compute the coatoms of st(5)
which are in st^\sa(5). They coincide with those coatoms that are in
sg^\sa(5) already.

In the meantime, CInet::Tools packages have been updated and there
is a noticeable increase in running time for every single program
(see the `git diff` for `timing.log`). I tried to quickly profile:

  * gen.pl: the only program which uses the heavily updated
    CInet::Propositional in any serious capacity; but it still
    finishes too fast to be the source of the slowdown;

  * coatoms.pl: turns out to be entirely I/O bound and doesn't
    use anything besides CInet::Base and CInet::ManySAT which
    were not updated outside of documentation.

I cannot give a definitive answer here as to what caused these
drastic slowdowns. It might be I/O on the host machine... But this
shows that the containerization does one of its jobs: verify that
answers stay the same and allow to track performance of the code
over time.